Refactors wet-bulb temperature plugins into separate module. Moves and renames test modules to match.
