+++ improver_tests/psychrometric_calculations/wet_bulb_temperature/test_WetBulbTemperature.py
@@ -38,9 +38,7 @@ from cf_units import Unit
 from iris.cube import Cube, CubeList
 from iris.tests import IrisTest
 
-from improver.psychrometric_calculations.psychrometric_calculations import (
-    WetBulbTemperature,
-)
+from improver.psychrometric_calculations.wet_bulb_temperature import WetBulbTemperature
 from improver.synthetic_data.set_up_test_cubes import set_up_variable_cube
